Adaptation of VolUniDirShell and VolCylShell jacobian transformations.
- VolUniDirShell : R is now referenced with respect to a plane which is
parralel to the principal 3D axes, but not necessarily passing trhough
the origin of the 3D plane.

- VolCylShell : Extension of the cylindrical jacobian transformation
along the X and Y axis, and not mandatorily along the Z axis. This new
Jacobian Type is valid only 3D, as it is for the Jacobian Type
VolUniDirShell.

- Some modifications may be required so that Get_Geometry.cpp becomes
more coherent to these Jacbian Types and avoid redundancies.