When exchanging the onelab parameters GetDP/1ResolutionChanges and GetDP/2PostOperationChoices, make sure to locally set them as *not* read-only.
Indeed, in case the parameter already exists in the DB and *is* read-only, we don't want to overwrite its value.

This should fix the models where e.g. "GetDP/1ResolutionChanges" is set as read-only in a pro file, and changed depending on another parameter.

This should have no impact on the "classical" way of writing multi-resolution models, where there is actually only a single resolution (typically "Analysis") that is parametrized with tests.